Issue #322: Move AwsKeyManager to its own module
