+++ credentials/pom.xml
@@ -108,6 +108,12 @@
       if desired.
     -->
     <dependency>
+      <groupId>xyz.block</groupId>
+      <artifactId>web5-keymanager-aws</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
       <groupId>io.ktor</groupId>
       <artifactId>ktor-client-mock-jvm</artifactId>
     </dependency>
