+++ keymanager-aws/pom.xml
@@ -13,9 +13,9 @@
   <modelVersion>4.0.0</modelVersion>
 
   <!-- Artifact Configuration -->
-  <artifactId>web5-crypto</artifactId>
-  <name>Web5 Crypto</name>
-  <description>Crypto libraries for Web5</description>
+  <artifactId>web5-keymanager-aws</artifactId>
+  <name>Web5 AWS KeyManager Implementation</name>
+  <description>AWS Implementation of KeyManager</description>
 
 
   <!-- Properties -->
@@ -31,7 +31,7 @@
     <!-- Internal Dependencies -->
     <dependency>
       <groupId>xyz.block</groupId>
-      <artifactId>web5-common</artifactId>
+      <artifactId>web5-crypto</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
@@ -43,31 +43,8 @@
 
     <!-- External -->
     <dependency>
-      <groupId>com.nimbusds</groupId>
-      <artifactId>nimbus-jose-jwt</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.module</groupId>
-      <artifactId>jackson-module-kotlin</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.google.crypto.tink</groupId>
-      <artifactId>tink</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.bouncycastle</groupId>
-      <artifactId>bcprov-jdk15to18</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.bouncycastle</groupId>
-      <artifactId>bcpkix-jdk15to18</artifactId>
-    </dependency>
-
-    <!-- Test Dependencies -->
-    <dependency>
-      <groupId>commons-codec</groupId>
-      <artifactId>commons-codec</artifactId>
-      <scope>test</scope>
+      <groupId>com.amazonaws</groupId>
+      <artifactId>aws-java-sdk-kms</artifactId>
     </dependency>
 
   </dependencies>
